Rest of media support providers refactored to match new API.
