+++ nima/observe/health/src/main/java/io/helidon/nima/observe/health/HealthFeature.java
@@ -30,7 +30,7 @@ import io.helidon.health.HealthCheck;
 import io.helidon.health.HealthCheckType;
 import io.helidon.health.spi.HealthCheckProvider;
 import io.helidon.nima.http.media.EntityWriter;
-import io.helidon.nima.http.media.jsonp.JsonpMediaSupportProvider;
+import io.helidon.nima.http.media.jsonp.JsonpSupport;
 import io.helidon.nima.servicecommon.HelidonFeatureSupport;
 import io.helidon.nima.webserver.http.HttpRules;
 import io.helidon.nima.webserver.http.HttpService;
@@ -109,7 +109,7 @@ public class HealthFeature extends HelidonFeatureSupport {
     }
 
     private void configureRoutes(HttpRules rules) {
-        EntityWriter<JsonObject> entityWriter = JsonpMediaSupportProvider.serverResponseWriter();
+        EntityWriter<JsonObject> entityWriter = JsonpSupport.serverResponseWriter();
 
         rules.get("/", new HealthHandler(entityWriter, details, all))
                 .get("/" + READINESS.defaultEndpoint(), new HealthHandler(entityWriter, details, ready))
