+++ nima/observe/health/src/main/java/io/helidon/nima/observe/health/SingleCheckHandler.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022 Oracle and/or its affiliates.
+ * Copyright (c) 2022, 2023 Oracle and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -28,7 +28,7 @@ import io.helidon.common.http.NotFoundException;
 import io.helidon.health.HealthCheck;
 import io.helidon.health.HealthCheckResponse;
 import io.helidon.nima.http.media.EntityWriter;
-import io.helidon.nima.http.media.jsonp.JsonpMediaSupportProvider;
+import io.helidon.nima.http.media.jsonp.JsonpSupport;
 import io.helidon.nima.webserver.http.Handler;
 import io.helidon.nima.webserver.http.ServerRequest;
 import io.helidon.nima.webserver.http.ServerResponse;
@@ -86,7 +86,7 @@ class SingleCheckHandler implements Handler {
 
         if (details) {
             try (OutputStream out = res.outputStream()) {
-                entityWriter.write(JsonpMediaSupportProvider.JSON_OBJECT_TYPE,
+                entityWriter.write(JsonpSupport.JSON_OBJECT_TYPE,
                                    HealthHelper.toJson(check.name(), response),
                                    out,
                                    req.headers(),
