+++ nima/tests/integration/sse/server/src/test/resources/META-INF/services/io.helidon.nima.http.media.spi.MediaSupportProvider
